requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.22.0
supabase==2.0.0
python-dotenv==1.0.1
numpy==1.26.4